Move ReferenceFactory to its own file
